requirements.txt
python-dotenv
feedparser
trafilatura
requests
beautifulsoup4
google-genai
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.0
